Registered user in user/admin: Registered a user in the Django admin interface.
Implemented user manager for create user: Implemented a custom user manager for creating users.

Installed django-debug-toolbar: Installed django-debug-toolbar for easier debugging and request monitoring.

Added endpoint to root urls for user and debugtoolbar: Added endpoints to the root URLs for user management and debugtoolbar.

Implemented views for Create and Login User: Implemented views for creating and logging in users.

Created serializers for user and made migrations: Created serializers for the user model and applied migrations to persist the models.